requirements.txt
psycopg2-binary==2.8.4
kafka-python==1.4.7
pymongo==3.10.0